Move initializeHints() to test
