+++ adapter/src/main/java/org/apache/calcite/prepare/KuduPrepareImpl.java
@@ -478,29 +478,4 @@ public class KuduPrepareImpl extends CalcitePrepareImpl {
     throw new RuntimeException("Unable to find KuduSchema in " + rootSchema);
   }
 
-  private static void setFinalStatic(Field field, Object newValue) throws Exception {
-    field.setAccessible(true);
-
-    Field modifiersField = Field.class.getDeclaredField("modifiers");
-    modifiersField.setAccessible(true);
-    modifiersField.setInt(field, field.getModifiers() & ~Modifier.FINAL);
-
-    field.set(null, newValue);
-  }
-
-  // TODO figure out a better way to set the HintStrategyTable for sql queries
-  public static void initializeHints() {
-    try {
-      SqlToRelConverter.Config CONFIG_MODIFIED = ImmutableBeans.create(SqlToRelConverter.Config.class)
-          .withRelBuilderFactory(RelFactories.LOGICAL_BUILDER)
-          .withRelBuilderConfigTransform(c -> c.withPushJoinCondition(true))
-          .withHintStrategyTable(KuduQuery.KUDU_HINT_STRATEGY_TABLE);
-      // change SqlToRelConverter.CONFIG to use one that has the above
-      // HintStrategyTable
-      setFinalStatic(SqlToRelConverter.class.getDeclaredField("CONFIG"), CONFIG_MODIFIED);
-    } catch (Exception e) {
-      throw new RuntimeException(e);
-    }
-  }
-
 }
