+++ adapter/src/test/java/com/twilio/kudu/sql/ScenarioIT.java
@@ -19,12 +19,17 @@ import com.twilio.kudu.dataloader.Scenario;
 import com.twilio.kudu.sql.schema.DefaultKuduSchemaFactory;
 import com.twilio.kudu.sql.schema.KuduSchema;
 import org.apache.calcite.prepare.KuduPrepareImpl;
+import org.apache.calcite.rel.core.RelFactories;
+import org.apache.calcite.sql2rel.SqlToRelConverter;
+import org.apache.calcite.util.ImmutableBeans;
 import org.apache.kudu.test.KuduTestHarness;
 import org.junit.BeforeClass;
 import org.junit.ClassRule;
 import org.junit.Test;
 
 import java.io.IOException;
+import java.lang.reflect.Field;
+import java.lang.reflect.Modifier;
 import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.PreparedStatement;
@@ -46,7 +51,7 @@ public class ScenarioIT {
 
   @BeforeClass
   public static void setup() throws SQLException, IOException {
-    KuduPrepareImpl.initializeHints();
+    initializeHints();
     String urlFormat = JDBCUtil.CALCITE_MODEL_TEMPLATE_DML_DDL_ENABLED + ";schema."
         + KuduSchema.CREATE_DUMMY_PARTITION_FLAG + "=false";
     JDBC_URL = String.format(urlFormat, DefaultKuduSchemaFactory.class.getName(),
@@ -73,6 +78,31 @@ public class ScenarioIT {
           .loadScenario(ScenarioIT.class.getResource("/scenarios/ReportCenter.UsageReportTransactions.json"));
       // load data
       new DataLoader(JDBC_URL, usageReportTransactionScenario).loadData(Optional.empty());
+    }
+  }
+
+  private static void setFinalStatic(Field field, Object newValue) throws Exception {
+    field.setAccessible(true);
+
+    Field modifiersField = Field.class.getDeclaredField("modifiers");
+    modifiersField.setAccessible(true);
+    modifiersField.setInt(field, field.getModifiers() & ~Modifier.FINAL);
+
+    field.set(null, newValue);
+  }
+
+  // TODO figure out a better way to set the HintStrategyTable for sql queries
+  public static void initializeHints() {
+    try {
+      SqlToRelConverter.Config CONFIG_MODIFIED = ImmutableBeans.create(SqlToRelConverter.Config.class)
+          .withRelBuilderFactory(RelFactories.LOGICAL_BUILDER)
+          .withRelBuilderConfigTransform(c -> c.withPushJoinCondition(true))
+          .withHintStrategyTable(KuduQuery.KUDU_HINT_STRATEGY_TABLE);
+      // change SqlToRelConverter.CONFIG to use one that has the above
+      // HintStrategyTable
+      setFinalStatic(SqlToRelConverter.class.getDeclaredField("CONFIG"), CONFIG_MODIFIED);
+    } catch (Exception e) {
+      throw new RuntimeException(e);
     }
   }
